Task_6/crud_app/requirements.txt
Django==5.2.9
djangorestframework
django-cors-headers
django-rest-passwordreset
djangorestframework-simplejwt
python-dotenv
Pillow
gunicorn>=21.2.0
psycopg2-binary
whitenoise



